Applied fixes to order services

+++ order-stock-microservices/order-service/src/main/java/org/orderservice/dto/product/ProductTransactionDTO.java
@@ -1,4 +1,4 @@
 package org.orderservice.dto.product;
 
-public record ProductTransactionDTO(Integer id, Integer amount){
+public record ProductTransactionDTO(Long id, Integer amount){
 }
